RealBankTransaction/requirements.txt
apache-flink==1.16.0
requests
openpyxl
pandas
kafka-python
kaggle
psycopg2-binary==2.9.1
requests


